Goal was to fix: Bug Report: Can't log in from profile page 🔗 Clone Bug Report: Can't log in from profile page 🔗 Steps to reproduce:
Load site with sample data populated.
Log in as sample / sosecret
Click on a username (e.g. AS) in a bloom shown in the timeline
Click logout
Log in as sample / sosecret
Expect to see: A logged in view of AS’s profile, with a “Logout” button in the top-right.

Actually see: A 501 error page saying “Server does not support this operation.”

:gear: bug

DONE now